requirements.txt
Flask
Flask-Cors
python-dotenv
requests
Pillow
openai
trafilatura
beautifulsoup4
